tests(semantics): add semantics tests and refactored test file structure to be more idomical Rust
KAG-2971
