+++ src/lib.rs
@@ -24,7 +24,6 @@ pub mod parser;
 pub mod router;
 pub mod schema;
 pub mod semantics;
-mod ast_tests;
 
 #[cfg(feature = "ffi")]
 pub mod ffi;
